DRY, moved redundant code to base classes
