+++ LibDmd/LibDmd.csproj
@@ -560,6 +560,7 @@
     </Compile>
     <Compile Include="Output\ZeDMD\ZeDMDHDWiFi.cs" />
     <Compile Include="Output\ZeDMD\ZeDMDHD.cs" />
+    <Compile Include="Output\ZeDMD\ZeDMDWiFiBase.cs" />
     <Compile Include="Output\ZeDMD\ZeDMDWiFi.cs" />
     <Compile Include="Output\ZeDMD\ZeDMDBase.cs" />
     <Compile Include="Output\ZeDMD\ZeDMD.cs" />
@@ -604,8 +605,8 @@
       <ItemGroup>
         <EmbeddedResource Include="Costura32\serum.dll" />
         <EmbeddedResource Include="Costura32\zedmd.dll" />
-		<EmbeddedResource Include="Costura32\libserialport.dll" />
-		<EmbeddedResource Include="Costura32\libSkiaSharp.dll" />
+        <EmbeddedResource Include="Costura32\libserialport.dll" />
+        <EmbeddedResource Include="Costura32\libSkiaSharp.dll" />
         <EmbeddedResource Include="Costura32\sqlite3.dll" />
         <EmbeddedResource Include="Costura32\ftd2xx.dll" />
       </ItemGroup>
@@ -616,8 +617,8 @@
       <ItemGroup>
         <EmbeddedResource Include="Costura64\serum64.dll" />
         <EmbeddedResource Include="Costura64\zedmd64.dll" />
-		<EmbeddedResource Include="Costura64\libserialport64.dll" />
-		<EmbeddedResource Include="Costura64\libSkiaSharp.dll" />
+        <EmbeddedResource Include="Costura64\libserialport64.dll" />
+        <EmbeddedResource Include="Costura64\libSkiaSharp.dll" />
         <EmbeddedResource Include="Costura64\sqlite3.dll" />
         <EmbeddedResource Include="Costura64\ftd2xx.dll" />
       </ItemGroup>
